+++ navigate/parser/_event.py
@@ -1,7 +1,7 @@
 # SPDX-FileCopyrightText: 2026 Fonden Mærsk Mc-Kinney Møller Center for Zero Carbon Shipping
 # SPDX-License-Identifier: Apache-2.0
 
-from navigate.parser._lark_parser import SourceLoc
+from navigate.parser._lark_parser import SourceLocation
 
 
 class Event:
@@ -9,22 +9,22 @@ class Event:
 
     Parameters
     ----------
-    source : SourceLoc
+    source : SourceLocation
         Location of the Date/Start keyword that created this event.
     deck_line : int
         Line in the .nav file of the enclosing INCLUDE directive.
     """
 
-    def __init__(self, source: SourceLoc, deck_line: int = 0):
-        self.stmts: list = []
+    def __init__(self, source: SourceLocation, deck_line: int = 0):
+        self.statements: list = []
         self._source = source
         self._deck_line = deck_line
 
-    def add_stmt(self, stmt):
-        self.stmts.append(stmt)
+    def add_statement(self, statement):
+        self.statements.append(statement)
 
     @property
-    def source(self) -> SourceLoc:
+    def source(self) -> SourceLocation:
         return self._source
 
     @property